API: split out MLIR-independent printing into BasePrinter and use in STIM
