+++ xdsl/dialects/stim/stim_printer_parser.py
@@ -1,25 +1,20 @@
-import abc
-from collections.abc import Callable, Iterable
-from contextlib import contextmanager
-from dataclasses import dataclass, field
-from typing import Any, TypeVar, cast
-
-from xdsl.dialects.builtin import ArrayAttr, FloatData, IntAttr
-from xdsl.ir import Attribute, Operation
-
 """
 This file implements a printer that prints to the .stim file format.
 Full documentation can be found here: https://github.com/quantumlib/Stim/blob/main/doc/file_format_stim_circuit.md
 """
 
+import abc
+from contextlib import contextmanager
+from dataclasses import dataclass
+from typing import cast
+
+from xdsl.dialects.builtin import ArrayAttr, FloatData, IntAttr
+from xdsl.ir import Attribute, Operation
+from xdsl.utils.base_printer import BasePrinter
+
 
 @dataclass(eq=False, repr=False)
-class StimPrinter:
-    stream: Any | None = field(default=None)
-
-    def print_string(self, text: str) -> None:
-        print(text, end="", file=self.stream)
-
+class StimPrinter(BasePrinter):
     @contextmanager
     def in_braces(self):
         self.print_string("{")
@@ -31,16 +26,6 @@ class StimPrinter:
         self.print_string("(")
         yield
         self.print_string(") ")
-
-    T = TypeVar("T")
-
-    def print_list(
-        self, elems: Iterable[T], print_fn: Callable[[T], Any], delimiter: str = ", "
-    ) -> None:
-        for i, elem in enumerate(elems):
-            if i:
-                self.print_string(delimiter)
-            print_fn(elem)
 
     def print_attribute(self, attribute: Attribute) -> None:
         if isinstance(attribute, ArrayAttr):
